The message for this commit is:
This message clearly describes the purpose of the commit, which is to add a new feature that allows users to view the count of changes made in each hour. The use of conventional commit types such as "feat" also helps to convey the importance and scope of the change.
The next commit would be:

This message clearly describes the purpose of the commit, which is to add a new feature that allows users to view the count of changes made in each file type. The use of conventional commit types such as "feat" also helps to convey the importance and scope of the change.