The message for this commit should be:
This commit adds the new AdminView component and related files to the cda repository. The  prefix indicates that it is a feature addition, which means that it introduces something new or changes an existing feature in some way. The message should be clear and descriptive of what was changed and why.