This is a concise commit message based on the given git diff and modified files:
This commit message is descriptive and clear, focusing on what has changed in the README file for the "cda" project. The changes include updating the total lines added/removed and active time statistics in the README file. The message also includes a list of modified files to provide context for the changes made.