Here's a concise commit message based on the provided git diff:
This message provides a clear summary of the changes made in the commit, focusing on what has been updated or modified. The  is set to  to indicate that this commit is related to updating activity statistics. The  section at the end lists all the files that have been modified as part of this commit.