The main changes in this commit are: * The total lines added and removed for the  project have been updated. * The  has increased by 2 minutes. * The  file has been modified, with 3 more lines of code being added and 0 lines being removed. * The  file has been updated to reflect the new changes in coding activity. Here is a concise commit message based on these changes:
In this message, we've used the  type to indicate that we are making changes to the codebase without adding new functionality or fixing bugs. We've also included a brief summary of the changes made in the commit, which includes updating the total lines added/removed and active time, as well as adding modifications to the  file.