Here's a possible commit message based on the given Git diff:
In this commit message, we're using the "refactor" type to indicate that we've made changes to the codebase without adding any new features or fixing any bugs. We've also included a brief description of the changes we've made, which is the optimization of the profile-hub.ts file.
The commit message includes the modified files, which helps other developers understand what has changed and why. The "Last Updated" timestamp at the end of the message indicates when the changes were made.