The commit message for this git diff could be:
In this message, we've used the conventional commit type "refactor" to indicate that the changes made were related to code refactoring rather than new features or bug fixes. We've also included a brief summary of the changes and the affected files in the commit message body. The "Last Updated" timestamp at the end of the message is not needed for this commit, as it is already included in the git diff output.