Here is a concise commit message based on the given git diff and modified files:
In this commit message, we use the conventional commit type "chore" to indicate that the changes made are related to housekeeping or maintenance tasks. The summary of the commit is "Update coding activity statistics for May 20, 2025." This includes the main changes and affected functionality, specifically updating the coding activity statistics for the date of May 20, 2025. We also provide a list of modified files that were affected by this change.