Here's a possible concise commit message based on the information provided:
This message describes the changes made to the  file, specifically updating the activation time from a previous value to the new value specified in the commit. It also lists the modified functions (only one in this case) and the files changed (one file). The type of commit is "refactor," indicating that it improves the codebase without adding new features or fixing bugs.