The commit message for this diff is:
This message indicates that the changes made to the project are related to the  feature and are focused on increasing the number of lines added and removed in the README file. The commit type is , which means "feature" or "new functionality". The scope is , indicating that the change is related to this specific feature. The summary provides a brief description of the changes made, highlighting the increase in the number of lines added and removed in the README file by 10%.