Here's a concise commit message based on the given git diff and modified files:
This message follows the conventional commit format by using the  type to indicate that we made changes to the codebase without adding any new features or fixing any bugs. We also provide a brief description of what we did, which is that we improved the efficiency of the skill tag modal component by reducing the number of lines changed and adding more active time. Finally, we list the modified files in the commit.