Here's a possible commit message based on the provided information:
This commit message follows the conventional commit format and uses the  type to indicate that we have made changes to the codebase without adding any new features or fixing any bugs. The message also includes a brief summary of the changes made, in this case updating the activity data for June 5th.
The modified files include:
*
*