Here is a possible commit message based on the provided diff:
In this commit message, we use the conventional type "fix" to indicate that we have made a change that fixes an existing issue or problem. We also include a brief description of the changes made, in this case, adjusting the total lines added and active time in the activity report for the cda project. This helps to provide context for the change and makes it easier for others to understand the purpose of the commit.