This is a concise commit message that summarizes the main changes and affected functionality in the commit. The message includes the following information: * Type of change (in this case, "refactor") * Scope of change (in this case, "cda") * Brief summary of changes * List of modified functions * List of files changed * Activity duration (X minutes) Here is an example commit message that incorporates the above information:
This commit message clearly describes the changes made in the commit, including the reason for the change (improving code readability) and the specific functions that were modified. It also includes a list of files changed and the activity duration, which helps developers understand the scope and impact of the commit.