This is a concise commit message based on the provided git diff and modified files:
This commit message uses the conventional commit type , which indicates that a bug or issue has been fixed. The scope of this commit is , as it relates to the CDA project.
The summary of this commit message is also clear and descriptive, stating that the lines added and removed in HelperGlossary.tsx have been adjusted due to formatting changes.
The modified files are listed at the end of the message, with  and .